Add IO object and statement/expression block
A statement/expression block is a set of statements/expressions
separated by semi-colons. The statements are executed in order and the
value of the block is value of the last expression.

i:IO declares that i is a variable of type IO. '<- new IO' allocates a new
object to the variable i.